requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.0
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.7.7
annotated-types==0.6.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==2.1.1
dataclasses-json==0.6.7
exceptiongroup==1.2.2
flake8==6.1.0
frozenlist==1.4.1
gigachain==0.2.16
gigachain-community==0.2.16.post3
gigachain-core==0.2.38.post2
gigachain-text-splitters==0.2.4
gigachat==0.1.35
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isort==5.13.2
jsonpatch==1.33
jsonpointer==3.0.0
langfuse==2.52.2
langsmith==0.1.137
magic-filter==1.0.12
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.10
packaging==24.1
pycodestyle==2.11.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.1.0
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==1.4.36
tenacity==8.5.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.3
wrapt==1.16.0
yarl==1.9.4
